requirements.txt
attrs==22.1.0
autopep8==1.7.0
backoff==2.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
flake8==5.0.4
Flask==2.2.2
Flask-Cors==3.0.10
googleapis-common-protos==1.56.4
grpcio==1.50.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
logzero==1.7.0
markup==0.2
MarkupSafe==2.1.1
mccabe==0.7.0
opentelemetry-api==1.13.0
opentelemetry-exporter-otlp-proto-grpc==1.13.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-proto==1.13.0
opentelemetry-sdk==1.13.0
opentelemetry-semantic-conventions==0.34b0
packaging==21.3
pluggy==1.0.0
prometheus-client==0.15.0
protobuf==3.20.3
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
PyYAML==6.0
requests==2.28.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
yattag==1.14.0
zipp==3.9.0
